requirements.txt
streamlit
openai
textblob
peewee
spacy
syllapy
scipy
scikit-learn
plotly
python-dotenv